Empty out MSNBTUtil, moving functions to IncompleteSburbCodeItem and GristHelper
